refactor: 인풋 관련 커스텀 훅은 useGame에서 분리
