+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./assets/css/modal.css";
-import { useMemo, useState } from "react";
+import { useMemo } from "react";
 import { ThemeProvider } from "styled-components";
 import {
   GameBoard,
@@ -12,15 +12,83 @@ import {
 import GlobalStyles from "./assets/styles/GlobalStyles";
 import useGame from "./hooks/useGame";
 import useModalMessage from "./hooks/useModalMessage";
+import useInput from "./hooks/useInput";
+import { GAME_STATUS, WORDS, WORD_LENGTH } from "./constants";
+import { currentMessage, getGameStatus, updateGuessEvaulations } from "./utils";
+
+const isWordInList = (word: string, words: string[]) => {
+  if (words.includes(word)) return true;
+  return false;
+};
+const isValidLength = (word: string, wordLength: number) => {
+  if (word.length !== wordLength) return false;
+  return true;
+};
 
 function App() {
-  const { state: gameState, actions: gameActions } = useGame();
+  const {
+    state: gameState,
+    actions: { updateGameStatus, resetGame },
+  } = useGame();
+  const {
+    state: inputState,
+    actions: { resetCurrentInput, clickDeleteButton, clickLetter },
+  } = useInput();
   const { modalMessages, addMessage, removeMessage } = useModalMessage();
-  // 검증함수를 App에서 정의, modal이 껴있음
-  // keyboard로 전달후 고차함수
-  // 커스텀 훅에서는 재료만 전달해서 최대한 재사용성을 확보하고
-  // 사용처에서는 여러 재료들을 조합, 원하는 동작을 만들어보자
-  // wordsEvaulated를 가지고 있어야 할까? 매번 계산하면 안될까?
+
+  const handleKeyUp = ({ buttonValue }: { buttonValue: string }) => {
+    if (buttonValue === "enter" || buttonValue === "Enter") {
+      const word = inputState.currentInput.join("");
+      if (!isValidLength(word, WORD_LENGTH)) {
+        // addMessage({ id: Date.now(), message: MESSAGE.WRONG_LENGTH });
+        return;
+      }
+
+      if (!isWordInList(word, WORDS)) {
+        // addMessage({ id: Date.now(), message: MESSAGE.WRONG_ANSWER });
+        return;
+      }
+
+      const updatedGuessesEvaulated = updateGuessEvaulations(
+        inputState.currentInput
+      )({
+        answer: gameState.answer,
+        turn: gameState.turn,
+        guessEvaulations: gameState.guessEvaulations,
+      });
+
+      const gameStatus = getGameStatus({
+        turn: gameState.turn,
+        guessEvaulations: updatedGuessesEvaulated,
+      });
+
+      const message = currentMessage({
+        turn: gameState.turn,
+        gameStatus: gameState.gameStatus,
+        answer: gameState.answer,
+      });
+      resetCurrentInput();
+      updateGameStatus({
+        guessEvaulations: updatedGuessesEvaulated,
+        gameStatus,
+        newGuess: inputState.currentInput.join(""),
+      });
+      if (gameStatus !== GAME_STATUS.DOING) {
+        // addMessage({ id: Date.now(), message });
+      }
+      return;
+    }
+
+    if (buttonValue === "<" || buttonValue === "Backspace") {
+      clickDeleteButton();
+      return;
+    }
+
+    if (/[^A-Za-z]|[A-Za-z]{2}/.test(buttonValue)) return;
+
+    clickLetter(buttonValue);
+    return;
+  };
   const isResultModalOpen = useMemo(() => {
     const isOpen = modalMessages.length === 0 && gameState.isGameComplete;
     return isOpen;
@@ -39,15 +107,12 @@ function App() {
         ))}
       </NavModalPortal>
       <div className="App">
-        <GameBoard {...gameState} />
-        <KeyBoard {...gameState} {...gameActions} />
+        <GameBoard {...gameState} {...inputState} />
+        <KeyBoard {...gameState} handleKeyUp={handleKeyUp} />
       </div>
       <ResultModalPortal>
         {isResultModalOpen && (
-          <ResultModal
-            {...gameState}
-            callback={() => gameActions.resetGame()}
-          />
+          <ResultModal {...gameState} callback={() => resetGame()} />
         )}
       </ResultModalPortal>
     </ThemeProvider>
